[NEW]: AI-Travel-Agent
This notebook showcases on deploying local LLM agents using the Langchain tools on Intel® Core™ Ultra Processors. The aim is to deploy an Travel Agent on the iGPU (integrated GPU) of the AIPC. For this, Llamacpp GPU backend is setup and the agent created using the local LLM model. The agent makes use of langchain toolkits and tools for travel related queries.

Provides a user friendly notebook and a user friendly streamlit app where end users can ask travel related questions which uses Langchain agents and utilizes maximum GPU utilization.